REF: Separate ptycho solvers by file in new module
Three different solvers each with submethods is too many
for one file
