+++ src/tike/ptycho/solvers/__init__.py
@@ -1,0 +1,8 @@
+from .admm import admm
+from .combined import combined
+from .divided import divided
+
+__all__ = [
+    "combined",
+    "divided",
+]
